Extract functions to keep it neat
